requirements.txt
boto3
psycopg2-binary==2.9.7
flask
flask_cors